Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > I believe README.android has to be updated. For instance, it has the version # and other meta information. 
> > > 
> > 
> > Yes. I should update README.android too. I will create another CL for this.
> 
> Thank you !
> 
> > > Moreover, shouldn't subset_noto_cjk.py be run to filter out some characters from the upstream font?
> > 
> > This patch already ran the subset_noto_cjk.py.
> 
> Great! It'd have been nice if you had explicitly mentioned that in the CL description. Otherwise, it's not clear. Anyway, when you update README.android, you can mention that in the CL description (even though it's already in README.android).

Last I heard the difference in size between the pre- and post-filtering of the characters from upstream was on the order of 1-200 bytes, which seems pretty small unless it's just the cmap that's being modified and no outlines are being removed.  Is there a test that confirms the unwanted characters are not in the fonts?

Patch-set: 3
Reviewer: Gerrit User 1003598 <1003598@85c56323-6fa9-3386-8a01-6480fb634889>