Update patch set 1
Patch Set 1: Code-Review+2

i thought for a second if this would work in a backward compatible manner and parse the previous fonts.

if the test here passes, it means that the last font we created a few months ago can be parsed; so less of a risk.

Patch-set: 1
Reviewer: Gerrit User 1106808 <1106808@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2