Fix subset_noto_cjk.py to run on Python3
One of the dependency of `noto-fonts/cjk/subset_noto_cjk.py`
now requires Python3. This patch fixes it to run on Python3.

A few changes in this patch:
1. Syntactic fixes for Python3.
2. `nototools.ttc_utils` requires a few fixes for Python3 and
   recent AFDKO, such as [PR536]. This patch adds an option
   to use `TTCollection` in `fonttools` instead.
3. Improved command line parsing by using `argparse` to
   support both engines, to ensure the transition does not
   cause regressions.

Locally verified that the script with up-to-date `nototools`
on Python3.8.5 can generate exactly the same files from Noto
CJK 2.001 except the timestamp and the checksum in the `head`
table.

[PR536]: https://github.com/googlefonts/nototools/pull/536

Change-Id: Ic2ebbb20d5344e70a5a48873410f00e681ad909f