Revert "Update NotoColorEmoji to support Emoji 15.1"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c6c91933-ecab-442c-9add-c5e071cfde86). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia9514becbb6508b3a8315e11b303f1bff42d7b4a