Revert "Update NotoColorEmoji to support Emoji 15.1"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d69bda4d-9f19-4758-8543-d6a0635393fc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I72ce881f08762946691255167c106be5cb6c77dd